requirements.txt
git+https://github.com/evennia/evennia.git#egg=evennia
boto3==1.12.36
psycopg2-binary==2.8.5
django-model-utils==4.0.0
django-bootstrap4==1.1.1
wagtail
coverage
coveralls